[#94] MemesController의 좋아요 관련 로직 분리하여 의존성을 낮춤
